+++ lib/chewy.rb
@@ -50,6 +50,7 @@ require 'chewy/repository'
 require 'chewy/runtime'
 require 'chewy/log_subscriber'
 require 'chewy/strategy'
+require 'chewy/clients'
 require 'chewy/index'
 require 'chewy/type'
 require 'chewy/fields/base'
@@ -141,15 +142,15 @@ module Chewy
       type
     end
 
-    # Main elasticsearch-ruby client instance
+    # Main client instance
     #
-    def client
-      Thread.current[:chewy_client] ||= begin
-        client_configuration = configuration.deep_dup
-        client_configuration.delete(:prefix) # used by Chewy, not relevant to Elasticsearch::Client
-        block = client_configuration[:transport_options].try(:delete, :proc)
-        ::Elasticsearch::Client.new(client_configuration, &block)
-      end
+    def client(name = nil)
+      ActiveSupport::Deprecation.warn "Method 'Chewy.client' is deprecated, use 'Chewy.client(name)' or 'Chewy.default_client' instead." if name.nil?
+      Chewy::Clients.with_name(name || :default)
+    end
+
+    def default_client
+      client(:default)
     end
 
     # Sends wait_for_status request to ElasticSearch with status
@@ -157,16 +158,20 @@ module Chewy
     #
     # Does nothing in case of config `wait_for_status` is undefined.
     #
-    def wait_for_status
-      client.cluster.health wait_for_status: Chewy.configuration[:wait_for_status] if Chewy.configuration[:wait_for_status].present?
+    def wait_for_status(name = :default)
+      client = client(name)
+      wait_for_status = clients[name][:wait_for_status]
+      client.cluster.health wait_for_status: wait_for_status if wait_for_status.present?
     end
 
     # Deletes all corresponding indexes with current prefix from ElasticSearch.
     # Be careful, if current prefix is blank, this will destroy all the indexes.
     #
     def massacre
-      Chewy.client.indices.delete(index: [Chewy.configuration[:prefix], '*'].reject(&:blank?).join('_'))
-      Chewy.wait_for_status
+      clients.each do |name, config|
+        client(name).indices.delete(index: [config[:prefix], '*'].reject(&:blank?).join('_'))
+        wait_for_status(name)
+      end
     end
     alias_method :delete_all, :massacre