+++ spec/chewy/index_spec.rb
@@ -53,6 +53,37 @@ describe Chewy::Index do
 
       specify { expect(Dummies1Index.client).to eq(Dummies2Index.client) }
     end
+
+    context 'when use alternative client_name' do
+      around do |example|
+        settings = Chewy.settings.dup
+        example.run
+        Chewy.settings = settings
+      end
+
+      before do
+        Chewy.settings = Chewy.settings.merge(
+          clients: {
+            dummy: {
+              hosts: 'localhost:9251'
+            }
+          }
+        )
+        stub_index(:dummies) do
+          use_client :dummy
+        end
+      end
+
+      specify do
+        expect(DummiesIndex.client).not_to eq(Chewy.default_client)
+        hosts = DummiesIndex.client.transport.hosts
+        expect(hosts.size).to eq 1
+
+        host = hosts.first
+        expect(host[:host]).to eq 'localhost'
+        expect(host[:port]).to eq 9251
+      end
+    end
   end
 
   describe '.index_name' do
@@ -67,7 +98,10 @@ describe Chewy::Index do
     specify { expect(stub_const('DevelopersIndex', Class.new(Chewy::Index)).index_name(prefix: 'test')).to eq('test_developers') }
 
     context do
-      before { allow(Chewy).to receive_messages(configuration: {prefix: 'testing'}) }
+      before do
+        Chewy.settings = Chewy.settings.merge(prefix: 'testing')
+      end
+
       specify { expect(DummiesIndex.index_name).to eq('testing_dummies') }
       specify { expect(stub_index(:dummies) { index_name :users }.index_name).to eq('testing_users') }
       specify { expect(stub_index(:dummies) { index_name :users }.index_name(prefix: '')).to eq('users') }
@@ -81,7 +115,7 @@ describe Chewy::Index do
   end
 
   describe '.prefix' do
-    before { allow(Chewy).to receive_messages(configuration: {prefix: 'testing'}) }
+    before { Chewy.settings = Chewy.clients[:default].merge(prefix: 'testing') }
     specify { expect(Class.new(Chewy::Index).prefix).to eq('testing') }
   end
 
@@ -162,9 +196,7 @@ describe Chewy::Index do
 
   describe '.settings' do
     before do
-      allow(Chewy).to receive_messages(config: Chewy::Config.send(:new))
-
-      Chewy.analyzer :name, filter: %w[lowercase icu_folding names_nysiis]
+      Chewy.analyzer :name, filter: %w(lowercase icu_folding names_nysiis)
       Chewy.analyzer :phone, tokenizer: 'ngram', char_filter: ['phone']
       Chewy.tokenizer :ngram, type: 'nGram', min_gram: 3, max_gram: 3
       Chewy.char_filter :phone, type: 'pattern_replace', pattern: '[^\d]', replacement: ''
@@ -182,7 +214,7 @@ describe Chewy::Index do
         tokenizer: {ngram: {type: 'nGram', min_gram: 3, max_gram: 3}},
         char_filter: {phone: {type: 'pattern_replace', pattern: '[^\d]', replacement: ''}},
         filter: {names_nysiis: {type: 'phonetic', encoder: 'nysiis', replace: false}}
-      }})
+      }, index: {number_of_shards: 1, number_of_replicas: 0}})
     end
   end
 
@@ -232,7 +264,16 @@ describe Chewy::Index do
   end
 
   describe '.settings_hash' do
-    before { allow(Chewy).to receive_messages(config: Chewy::Config.send(:new)) }
+    specify do
+      expect(stub_index(:documents).settings_hash).to eq(
+        settings: {
+          index: {
+            number_of_shards: 1,
+            number_of_replicas: 0
+          }
+        }
+      )
+    end
 
     specify { expect(stub_index(:documents).settings_hash).to eq({}) }
     specify { expect(stub_index(:documents) { settings number_of_shards: 1 }.settings_hash).to eq(settings: {number_of_shards: 1}) }
@@ -272,6 +313,33 @@ describe Chewy::Index do
                end
              end.specification_hash.keys).to eq([:mappings])
     end
+  end
+
+  describe '.index_params' do
+    specify do
+      expect(stub_index(:documents).index_params).to eq(
+        settings: {
+          index: {
+            number_of_shards: 1,
+            number_of_replicas: 0
+          }
+        }
+      )
+    end
+
+    specify do
+      expect(stub_index(:documents) { settings number_of_shards: 1 }.index_params.keys).to eq([:settings])
+    end
+
+    specify do
+      index = stub_index(:documents) do
+        define_type :document do
+          field :name, type: 'string'
+        end
+      end
+      expect(index.index_params.keys).to eq([:settings, :mappings])
+    end
+
     specify do
       expect(stub_index(:documents) do
                settings number_of_shards: 1